Separate model and ribosome protein dynamics
