+++ feature-evolution/src/commonMain/kotlin/org/pointyware/ember/evolution/ModelProteinDynamics.kt
@@ -3,14 +3,6 @@ package org.pointyware.ember.evolution
 import org.pointyware.ember.entities.networks.Network
 
 /**
- * Folders abstract away all the chemical and mechanical details of protein folding
- * for their respective dynamics.
- */
-interface Folder<Output> {
-    fun fold(polypeptides: List<List<AminoAcid>>): Output
-}
-
-/**
  *
  */
 class ModelProteinDynamics: Folder<Network> {
@@ -23,13 +15,5 @@ class ModelProteinDynamics: Folder<Network> {
         val blockProteins = polypeptides.map {
 
         }
-
-
-    }
-}
-
-class RibosomeProteinDynamics: Folder<Ribosome> {
-    override fun fold(polypeptides: List<List<AminoAcid>>): Ribosome {
-        TODO("Implement")
     }
 }
