+++ feature-evolution/src/commonMain/kotlin/org/pointyware/ember/evolution/RibosomeProteinDynamics.kt
@@ -1,0 +1,7 @@
+package org.pointyware.ember.evolution
+
+class RibosomeProteinDynamics: Folder<Ribosome> {
+    override fun fold(polypeptides: List<List<AminoAcid>>): Ribosome {
+        TODO("Implement")
+    }
+}
